[CITE-163] Separated parseMetaDataNote & parseExtra from citationFactory
